update release numbers and move tests of depreactions to a separate test module
