+++ scipy/sparse/_lil.py
@@ -528,7 +528,7 @@ def isspmatrix_lil(x):
         is elementwise multiplication. See the spmatrix to sparray migration guide
         https://docs.scipy.org/doc/scipy/reference/sparse.migration_to_sparray.html
 
-        The isspmatrix_lil function will be removed no earlier than v2.1.
+        The isspmatrix_lil function will be removed no earlier than v2.2.
         """
     prefixes = (os.path.dirname(__file__),)
     warn(msg, category=DeprecationWarning, skip_file_prefixes=prefixes)
