+++ scipy/sparse/_matrix.py
@@ -27,7 +27,7 @@ class spmatrix:
         is elementwise multiplication. See the spmatrix to sparray migration guide
         https://docs.scipy.org/doc/scipy/reference/sparse.migration_to_sparray.html
 
-        The spmatrix classes will be removed no earlier than v2.1.
+        The spmatrix classes will be removed no earlier than v2.2.
         """
         prefixes = (os.path.dirname(__file__),)
         warn(msg, category=DeprecationWarning, skip_file_prefixes=prefixes)
